+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Generic/PrettyCore.hs
@@ -1288,7 +1288,7 @@ ppBabbageUtxoPredFailure (IncorrectTotalCollateralField c1 c2) =
 ppBabbageUtxoPredFailure (BabbageOutputTooSmallUTxO xs) =
   ppSexp "BabbageOutputTooSmallUTxO" [ppList (ppPair (pcTxOut reify) pcCoin) xs]
 ppBabbageUtxoPredFailure (BabbageNonDisjointRefInputs xs) =
-  ppSexp "BabbageNonDisjointRefInputs" [ppSet pcTxIn xs]
+  ppSexp "BabbageNonDisjointRefInputs" [ppList pcTxIn (toList xs)]
 
 instance Reflect era => PrettyA (BabbageUtxoPredFailure era) where
   prettyA = ppBabbageUtxoPredFailure